ref111

+++ php/actions/blacklist.php
@@ -9,30 +9,25 @@ try {
         throw new Exception();
     }
 
-    parse_str($_POST['topics_ids']);
+    parse_str($_POST['topics_ids'], $data);
 
     $value = empty($_POST['value']) ? 0 : 1;
 
-    $topics_ids = array_chunk($topics_ids, 500);
+    $data['topics_ids'] = array_chunk($data['topics_ids'], 500);
 
-    foreach ($topics_ids as $topics_ids) {
-        switch ($value) {
-            case 0:
-                $in = str_repeat('?,', count($topics_ids) - 1) . '?';
-                Db::query_database(
-                    "DELETE FROM Blacklist WHERE id IN ($in)",
-                    $topics_ids
-                );
-                break;
-            case 1:
-                $select = str_repeat('SELECT ? UNION ALL ', count($topics_ids) - 1) . ' SELECT ?';
-                Db::query_database(
-                    "INSERT INTO Blacklist (id) $select",
-                    $topics_ids
-                );
-                break;
-            default:
-                throw new Exception("Error: Неизвестное событие");
+    foreach ($data['topics_ids'] as $topics_ids) {
+        if ($value == 0) {
+            $in = str_repeat('?,', count($topics_ids) - 1) . '?';
+            Db::query_database(
+                "DELETE FROM Blacklist WHERE id IN ($in)",
+                $topics_ids
+            );
+        } elseif ($value == 1) {
+            $select = str_repeat('SELECT ? UNION ALL ', count($topics_ids) - 1) . ' SELECT ?';
+            Db::query_database(
+                "INSERT INTO Blacklist (id) $select",
+                $topics_ids
+            );
         }
     }
